refactor(codegen): extract lower_closure_check into locus/closure.rs (Round 4e)
1 fn lifted into LocusClosure<'ctx>: lower_closure_check (~601 lines).
This is the closure-assertion lowering called from inside each of the
synthesized __<epoch>_closures bodies — substitutes accumulator slots
for sum() / count() / mean() references, lowers the
`left == right within tolerance` assertion, and routes violations to
the matching on_failure handler (or dprintf+exit when none declared).

Surface bumps to pub(crate):
- Cx fields: accumulator_ctx
- Cx methods: update_accumulator_slot
- Structs: AccumulatorCtx (+ fields)
- Free fn: i128_const

locus/method.rs now imports LocusClosure so the Phase-B method
bodies' synthesized __<epoch>_closures fn-body emission keeps
resolving.

codegen.rs: 28683 -> 28078 (-605 lines).
