+++ crates/hale-codegen/src/locus/mod.rs
@@ -1,6 +1,7 @@
 //! Locus codegen (lifecycle, methods, dissolve, arenas, placement,
 //! return path). Round 4 of the codegen model-org refactor.
 
+pub(crate) mod closure;
 pub(crate) mod decl;
 pub(crate) mod dissolve;
 pub(crate) mod instantiation;
